Purge references to `async-io` gem.
