+++ test/falcon/rackup/handler/falcon.rb
@@ -9,9 +9,8 @@ require 'falcon/server'
 require 'async/http/endpoint'
 require 'async/http/client'
 require 'async/process'
-require 'rackup/handler'
 
-require 'rack/handler/falcon'
+require 'rackup/handler/falcon'
 
 RackupHandler = Sus::Shared("rackup handler") do |server_name|
 	include Sus::Fixtures::Async::ReactorContext
@@ -49,7 +48,7 @@ RackupHandler = Sus::Shared("rackup handler") do |server_name|
 	end
 end
 
-describe Rackup::Handler::Falcon do
+describe Falcon::Rackup::Handler do
 	it_behaves_like RackupHandler, 'falcon'
 
 	let(:app) {lambda {|env| [200, {}, ["Hello World"]]}}
